fix(messages): `Showcmd` messages no longer update statuscolumn
This is to prevent visual artifacts from appearing when doing things
like `v<Right>ff<Esc>`.

refactor(messages): Changed how the cursor is centered in the `shocmd`
window

This is to prevent constantly updating the `sidescrolloff` command.

The cursor is placed on 1/2 of the display width of the window, instead
of the byte width.